sandbox/requirements.txt
# requirements.txt
nautilus_trader[all]>=1.198.0
python-dotenv>=1.0.0
redis>=4.5.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
requests>=2.28.0
websockets>=11.0
